Be more strict about multi-line strings and parens.
For a multiline string, the open quote must be at the end of the line,
and the close quote must be at the start of the line.  All whitespace at
the start of the string is stripped.  The first non-whitespace character
determines the indentation of the multiline string; subsequent lines
must have greater or equal indent and will have the leading spaces up to
the same indent stripped.

All parens must start and end at the same indentation level (from the
start of the line containing the open paren, not the paren itself); any
expression inside the parens must be indented relative to the line with
the open paren.